# docs/openapi.yaml
openapi: 3.0.3
info:
  title: go-mall API
  version: 0.1.0
  description: |
    go-mall 单商户商城 MVP 接口文档。

    当前文档作为后端、小程序前端、网页前端共同使用的接口契约。
    金额字段统一使用分，避免浮点数精度问题。

    标注为“规划接口，后端尚未实现”的接口用于前端 Mock 和后续后端开发排期。
servers:
  - url: http://localhost:8080
    description: local
...
paths:
...
  /api/v1/products:
    get:
      tags:
        - Product
      summary: 商品列表
      operationId: listProducts
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 页码
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
          description: 每页数量
        - name: category_id
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: 分类 ID
        - name: keyword
          in: query
          required: false
          schema:
            type: string
          description: 搜索关键词
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProductPage"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/auth/refresh:
    post:
      tags:
        - Auth
      summary: 刷新访问令牌
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: 刷新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/uploads:
    post:
      tags:
        - Upload
      summary: 上传图片
      description: |
        已实现。服务端根据配置上传到腾讯云 COS 或七牛 Kodo。
        默认最大 10 MB，仅支持 JPEG、PNG、WebP 和 GIF；具体上限由后端配置决定。
        返回的 URL 可用于头像、商品图片和售后凭证等业务字段。
      operationId: uploadImage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                scene:
                  type: string
                  pattern: "^[a-z0-9_-]{1,32}$"
                  default: general
                  description: 业务场景，用于组织对象目录，如 avatar、product、after_sale。
                  example: product
      responses:
        "200":
          description: 上传成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UploadImage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    merchantBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Merchant JWT
...
  responses:
...
    BadRequest:
      description: 参数错误或业务校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 40000
            message: 参数错误
    Unauthorized:
      description: 未登录或 token 无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 401
            message: 未登录
...
    InternalError:
      description: 服务内部错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 50000
            message: 服务内部错误
  schemas:
    SuccessResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          example: 40000
        message:
          type: string
          example: 参数错误
...
    ProductListItem:
      type: object
      required:
        - id
        - merchant_id
        - category_id
        - name
        - cover
        - min_price
      properties:
        id:
          type: integer
          format: int64
          example: 1
        merchant_id:
          type: integer
          format: int64
          example: 1
        category_id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: iPhone 15
        cover:
          type: string
          example: https://example.com/iphone.jpg
        min_price:
          type: integer
          format: int64
          description: 最低 SKU 价格，单位分
          example: 599900
    ProductPage:
      type: object
      required:
        - list
        - page
        - page_size
        - total
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/ProductListItem"
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 10
        total:
          type: integer
          format: int64
          example: 1
...
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          example: 7a2c1f80-0000-0000-0000-000000000000
    AuthResult:
      type: object
      required:
        - access_token
        - refresh_token
        - user
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          example: 7a2c1f80-0000-0000-0000-000000000000
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required:
        - id
        - nickname
        - avatar
        - mobile
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nickname:
          type: string
          example: 测试用户
        avatar:
          type: string
          example: https://example.com/avatar.jpg
        mobile:
          type: string
          example: "13800138000"
        gender:
          type: string
          example: 保密
        birthday:
          type: string
          example: "1995-01-01"
        bio:
          type: string
          example: 这个人很懒，还没有填写简介
...
    UploadImage:
      type: object
      required:
        - url
        - key
        - filename
        - content_type
        - size
        - provider
      properties:
        url:
          type: string
          example: https://images.example.com/go-mall/product/2026/07/550e8400-e29b-41d4-a716-446655440000.jpg
        key:
          type: string
          example: go-mall/product/2026/07/550e8400-e29b-41d4-a716-446655440000.jpg
        filename:
          type: string
          example: product.jpg
        content_type:
          type: string
          example: image/jpeg
        size:
          type: integer
          format: int64
          description: 文件大小，单位字节
          example: 102400
        provider:
          type: string
          enum:
            - cos
            - qiniu
          example: cos